classic logic organized and fetch logic moved
